qemu-setup: A role to setup QEMU and libvirt.
Set up for virtualization by installing QEMU and libvirt. For now we
only install the packaged version of QEMU. Later we may want to
support building and installing more recent releases.

We also add namespacing for variables and a Projects folder for our
repos since this is the way we are trending for this stuff.