vm_create: Role for creating VMs on target(s)
This role can be used to create Virtual Machines (VMs) on remote
target(s) and is based on [1]. Note this role depends on the
qemu_setup role.

Note that we use the virt-ubuntu-install script from the qemu-minimal
repo. See the README.md for more information on how to run this role.

[1]: https://www.redhat.com/sysadmin/build-VM-fast-ansible

Fixes #51.