docker_setup: Add a role for installing docker
We use an existing ansible-galaxy role [1] to install docker on the
target(s). Also sets up the user as a docker user. Note we install
docker-ce and docker-compose. We also need to update the GitHub
workflow so we pull in the role from Ansible Galaxy. We also need to
gather facts and we should really be doing this anyway so change this
to true.

Oh and we fix an issue with the roles_path in ansible_cfg.

[1]: https://galaxy.ansible.com/geerlingguy/docker

Fixes #57.