qemu_setup: Add support for bridge creation libvirt networking.
Update the qemu_setup role to add the optional creation of a network
bridge on the target(s). This allows for VMs to either be added to the
default NAT or to the bridge. When added to the bridge the VM appears
as a new host on the LAN and can be accessed as such.

Resolves #70.