consul_setup: Add a role to setup consul
We add a role that installs consul and then configures it correctly
for my home lab. We use ansible-vault secrets where necessary to
protect the gossip and TLS keys.

We also ensure prometheus-node-exporter is installed and register this
as a service via another config file. This then ensures the new node
is automagically added to our Prometheus+Grafana setup.

Resolves #68.