Ubuntu Jammy: Require that the host(s) are running a specific Ubuntu release
For now we lock down the OS disribution and release on all host(s) and
all roles to Ubuntu 22.04 LTS (code-name jammy). This simplifies
several parts of our scripting and is a totally acceptable requirement
(for now).

In order to do this we introduce a new role called check_platform
which compares the host(s) OS and release version against what is
contained in the meta/main.yml platform section of each role.

At the same time we clean up the permitted version of ansible on the
Ansible controller node via a requirements.txt file. We also clean up
ansible-lint to support the new YAML. We update the README.md
accordingly.

For CI purposes we add a test for Ubuntu 20.04.

Fixes #69.
Allows us to close #61 and #4.